Refactor of generation functions
Breaking up the code for generating elements to make it a biut easier to rtead and understand and also added some configuration options in the console before the simulation runs.
