+++ config.py
@@ -1,0 +1,22 @@
+# I used this dictionary of english-sounding words / syllables to generate the names of the depots
+names = {
+        "prefix" : [
+            "abing", "aber", "bark", "berry", "chelms", "car", "wood","strem", "ling", "spack",
+            "trem", "mug", "deck", "tillin", "nod", "mabb", "golling", "miad", "ebbing", "nike",
+            "dell"
+    ],
+        "suffix" : [
+            "shire", "ton", "don", "ford", "cester", "ley", "e"
+    ]}
+
+ukMap = [
+    [20, 10], [30, 10], 
+    [20, 20], [30, 20], 
+    [20, 30], [30, 30], 
+    [20, 40], [30, 40], 
+    [30, 50], [40, 50], 
+    [20, 60], [30, 60], [40, 60], 
+    [10, 70], [20, 70], [30, 70], [40, 70], [ 50, 70], 
+    [20, 80], [30, 80], [40, 80], [50, 80], 
+    [10, 90], [20, 90], [30, 90], [40, 90]
+]
